requirements.txt
Appium-Python-Client==0.50
behave==1.2.6
certifi==2019.11.28
chardet==3.0.4
idna==2.7
Jinja2==2.11.1
MarkupSafe==1.1.1
nose==1.3.7
nose-htmloutput==0.6.0
nose-parameterized==0.6.0
nose-testconfig==0.10
nose-xunitmp==0.4.0
parse==1.15.0
parse-type==0.5.2
requests==2.20.1
selenium==3.141.0
six==1.14.0
urllib3==1.24.3
